kiwi: Use common CMHW gestures
* Make use of CMHW gestures
* Refactor cmactions to Doze
* Integrate doze with proper settings (in display)

* Contains work of:
Patrick Lower <devvortex@gmail.com>
Arne Coucheron <arco68@gmail.com>
Christopher R. Palmer <crpalmer@gmail.com>
Dave Daynard <nardholio@gmail.com>
Brandon McAnsh <brandon.mcansh@gmail.com>
TheStrix <parthbhatia98@gmail.com>
Steve Kondik <steve@cyngn.com>
Altaf-Mahdi <altaf.mahdi@gmail.com>
Zhao Wei Liew <zhaoweiliew@gmail.com>
