--- cmactions/AndroidManifest.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.cyanogenmod.settings.device"
-    android:versionCode="1"
-    android:versionName="1.0"
-    android:sharedUserId="android.uid.system"
-    >
-
-    <uses-permission android:name="android.permission.DEVICE_POWER" />
-    <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
-    <uses-permission android:name="android.permission.VIBRATE" />
-    <uses-permission android:name="android.permission.WAKE_LOCK" />
-
-    <protected-broadcast android:name="com.android.systemui.doze.pulse" />
-
-    <uses-sdk
-        android:minSdkVersion="23"
-        android:targetSdkVersion="23" />
-
-    <application
-        android:allowBackup="true"
-        android:label="@string/device_settings_app_name"
-        android:persistent="true"
-        android:icon="@mipmap/ic_launcher_settings"
-        android:theme="@style/CMActions" >
-        <receiver android:name="com.cyanogenmod.settings.device.BootCompletedReceiver">
-            <intent-filter>
-                <action android:name="android.intent.action.BOOT_COMPLETED" />
-                <category android:name="android.intent.category.DEFAULT" />
-            </intent-filter>
-        </receiver>
-
-        <service android:name="ServiceWrapper" android:permission="ServiceWrapper">
-            <intent-filter>
-                <action android:name="ServiceWrapper"/>
-            </intent-filter>
-        </service>
-
-        <service android:name="com.cyanogenmod.settings.device.SensorsDozeService"
-                 android:permission="SensorsDozeService">
-        </service>
-
-        <service android:name="com.cyanogenmod.settings.device.SensorsDozeService"
-                 android:permission="SensorsDozeService">
-        </service>
-
-        <!-- stub to ensure its loaded - DO NOT REMOVE -->
-        <activity android:name=".KeyHandler"/>
-
-        <activity
-                  android:excludeFromRecents="true"
-                  android:name="CMActions"
-                  android:launchMode="singleInstance"
-                  android:screenOrientation="portrait"
-                  android:windowAnimationStyle="@null">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-            </intent-filter>
-        </activity>
-
-        <activity
-            android:name=".TouchscreenGestureActivity"
-            android:label="@string/screen_gestures_panel_title">
-            <intent-filter>
-                <action android:name="com.android.settings.action.EXTRA_SETTINGS" />
-            </intent-filter>
-            <meta-data
-                android:name="com.android.settings.category"
-                android:value="com.android.settings.category.device" />
-            <meta-data
-                android:name="com.android.settings.icon"
-                android:resource="@drawable/ic_settings_gestures" />
-        </activity>
-
-    </application>
-
-</manifest>
