--- cmactions/proguard.flags
@@ -1,9 +1,0 @@
--keep class org.cyanogenmod.CMActions.* {
-  *;
-}
-
--keepclasseswithmembers class * {
-    public <init>(android.content.Context, android.util.AttributeSet);
-}
-
--keep class ** extends android.support.v14.preference.PreferenceFragment
