--- cmactions/res/xml/touchscreen_panel.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-     Copyright (C) 2015 The CyanogenMod Project
-
-     Licensed under the Apache License, Version 2.0 (the "License"
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <PreferenceCategory
-        android:title="@string/gesture_display_title">
-
-        <SwitchPreference
-            android:key="touchscreen_gesture_c"
-            android:title="@string/touchscreen_c_gesture_title"
-            android:summary="@string/touchscreen_c_gesture_summary" />
-
-        <SwitchPreference
-            android:key="touchscreen_gesture_e"
-            android:title="@string/touchscreen_e_gesture_title"
-            android:summary="@string/touchscreen_e_gesture_summary" />
-
-        <SwitchPreference
-            android:key="touchscreen_gesture_m"
-            android:title="@string/touchscreen_m_gesture_title"
-            android:summary="@string/touchscreen_m_gesture_summary" />
-
-        <SwitchPreference
-            android:key="touchscreen_gesture_w"
-            android:title="@string/touchscreen_w_gesture_title"
-            android:summary="@string/touchscreen_w_gesture_summary" />
-
-    </PreferenceCategory>
-
-    <PreferenceCategory
-        android:key="ambient_display_key"
-        android:title="@string/ambient_display_title" >
-
-        <SwitchPreference
-            android:key="ambient_display_enable"
-            android:defaultValue="true"
-            android:title="@string/ambient_display_enable_title"
-            android:summary="@string/ambient_display_enable_summary" />
-
-        <SwitchPreference
-            android:key="gesture_hand_wave"
-            android:defaultValue="false"
-            android:title="@string/hand_wave_gesture_title"
-            android:summary="@string/hand_wave_gesture_summary" />
-
-        <SwitchPreference
-            android:key="gesture_pick_up"
-            android:defaultValue="false"
-            android:title="@string/pick_up_gesture_title"
-            android:summary="@string/pick_up_gesture_summary" />
-
-        <SwitchPreference
-            android:key="gesture_pocket"
-            android:defaultValue="false"
-            android:title="@string/pocket_gesture_title"
-            android:summary="@string/pocket_gesture_summary" />
-
-    </PreferenceCategory>
-
-    <PreferenceCategory
-        android:title="@string/touchscreen_gesture_extras">
-
-        <SwitchPreference
-            android:key="touchscreen_gesture_haptic_feedback"
-            android:persistent="false"
-            android:title="@string/touchscreen_gesture_haptic_feedback"
-            android:summary="@string/touchscreen_gesture_haptic_feedback_summary" />
-
-    </PreferenceCategory>
-
-</PreferenceScreen>
