--- cmactions/src/com/cyanogenmod/settings/device/BootCompletedReceiver.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2015 The CyanogenMod Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.cyanogenmod.settings.device;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.util.Log;
-
-public class BootCompletedReceiver extends BroadcastReceiver {
-    static final String TAG = "CMActions";
-
-    @Override
-    public void onReceive(final Context context, Intent intent) {
-        Log.d(TAG, "Booting");
-        context.startService(new Intent(context, SensorsDozeService.class));
-        // Call here to have the sysnode written / restored
-        CMActionsSettings.updateGestureMode(context);
-    }
-}
