--- cmactions/src/com/cyanogenmod/settings/device/TouchscreenGestureActivity.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2016 The CyanogenMod Project
- * Copyright (C) 2017 The LineageOS Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.cyanogenmod.settings.device;
-
-import android.os.Bundle;
-
-import com.android.settingslib.drawer.SettingsDrawerActivity;
-
-public class TouchscreenGestureActivity extends SettingsDrawerActivity {
-
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        getFragmentManager().beginTransaction().replace(R.id.content_frame,
-                new TouchscreenGestureFragment()).commit();
-    }
-}
