+++ doze/AndroidManifest.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.cyanogenmod.settings.doze"
+    android:versionCode="1"
+    android:versionName="1.0"
+    android:sharedUserId="android.uid.system"
+    >
+
+    <uses-permission android:name="android.permission.DEVICE_POWER" />
+    <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
+
+    <protected-broadcast android:name="com.android.systemui.doze.pulse" />
+
+    <uses-sdk
+        android:minSdkVersion="23"
+        android:targetSdkVersion="23" />
+
+    <application
+        android:allowBackup="true"
+        android:persistent="true" >
+
+        <receiver android:name="com.cyanogenmod.settings.doze.BootCompletedReceiver">
+            <intent-filter>
+                <action android:name="android.intent.action.BOOT_COMPLETED" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </receiver>
+
+        <service android:name="com.cyanogenmod.settings.doze.SensorsDozeService"
+                 android:permission="SensorsDozeService">
+        </service>
+
+        <activity
+            android:name=".DozeSettings"
+            android:label="@string/ambient_display_title"
+            android:theme="@style/DozeSettings">
+            <intent-filter>
+                <action android:name="org.cyanogenmod.settings.device.DOZE_SETTINGS" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+            <meta-data
+                android:name="org.cyanogenmod.settings.summary.receiver"
+                android:value="com.cyanogenmod.settings.doze.DozeReceiver" />
+            <meta-data
+                android:name="org.cyanogenmod.settings.summary.key"
+                android:value="doze_device_settings" />
+        </activity>
+
+    </application>
+
+</manifest>
