+++ doze/res/values/dimens.xml
