+++ doze/src/com/cyanogenmod/settings/doze/DozeSettingsFragment.java
@@ -1,0 +1,141 @@
+/*
+ * Copyright (C) 2015-2016 The CyanogenMod Project
+ * Copyright (C) 2017 The LineageOS Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.cyanogenmod.settings.doze;
+
+import android.app.Activity;
+import android.app.AlertDialog;
+import android.app.Dialog;
+import android.app.DialogFragment;
+import android.content.DialogInterface;
+import android.content.SharedPreferences;
+import android.os.Bundle;
+import android.provider.Settings;
+import android.support.v14.preference.PreferenceFragment;
+import android.support.v14.preference.SwitchPreference;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.CompoundButton;
+import android.widget.Switch;
+
+public class DozeSettingsFragment extends PreferenceFragment implements
+        CompoundButton.OnCheckedChangeListener {
+
+    private static final String KEY_GESTURE_HAND_WAVE = "gesture_hand_wave";
+    private static final String KEY_GESTURE_PICK_UP = "gesture_pick_up";
+    private static final String KEY_GESTURE_POCKET = "gesture_pocket";
+
+    private Switch mSwitch;
+    private SwitchPreference mHandwavePreference;
+    private SwitchPreference mPickupPreference;
+    private SwitchPreference mPocketPreference;
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        addPreferencesFromResource(R.xml.doze_settings);
+
+        SharedPreferences prefs = getActivity().getSharedPreferences("doze_settings",
+                Activity.MODE_PRIVATE);
+        if (savedInstanceState == null && !prefs.getBoolean("first_help_shown", false)) {
+            showHelp();
+        }
+
+        boolean dozeEnabled = isDozeEnabled();
+
+        mHandwavePreference = (SwitchPreference) findPreference(KEY_GESTURE_HAND_WAVE);
+        mHandwavePreference.setEnabled(dozeEnabled);
+        mPickupPreference = (SwitchPreference) findPreference(KEY_GESTURE_PICK_UP);
+        mPickupPreference.setEnabled(dozeEnabled);
+        mPocketPreference = (SwitchPreference) findPreference(KEY_GESTURE_POCKET);
+        mPocketPreference.setEnabled(dozeEnabled);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        final View view = LayoutInflater.from(getContext()).inflate(R.layout.doze, container, false);
+        ((ViewGroup) view).addView(super.onCreateView(inflater, container, savedInstanceState));
+               return view;
+    }
+
+    @Override
+    public void onViewCreated(View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        View switchBar = view.findViewById(R.id.switch_bar);
+        mSwitch = (Switch) switchBar.findViewById(android.R.id.switch_widget);
+        mSwitch.setChecked(isDozeEnabled());
+        mSwitch.setOnCheckedChangeListener(this);
+
+        switchBar.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mSwitch.setChecked(!mSwitch.isChecked());
+            }
+        });
+    }
+
+    private boolean enableDoze(boolean enable) {
+        return Settings.Secure.putInt(getContext().getContentResolver(),
+                Settings.Secure.DOZE_ENABLED, enable ? 1 : 0);
+    }
+
+    private boolean isDozeEnabled() {
+        return Settings.Secure.getInt(getContext().getContentResolver(),
+                Settings.Secure.DOZE_ENABLED, 1) != 0;
+    }
+
+    @Override
+    public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
+        boolean ret = enableDoze(b);
+        if (ret) {
+            mHandwavePreference.setEnabled(b);
+            mPickupPreference.setEnabled(b);
+            mPocketPreference.setEnabled(b);
+        }
+    }
+
+    public static class HelpDialogFragment extends DialogFragment {
+        @Override
+        public Dialog onCreateDialog(Bundle savedInstanceState) {
+            return new AlertDialog.Builder(getActivity())
+                    .setTitle(R.string.doze_settings_help_title)
+                    .setMessage(R.string.doze_settings_help_text)
+                    .setNegativeButton(R.string.dialog_ok, new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            dialog.cancel();
+                        }
+                    })
+                    .create();
+        }
+
+        @Override
+        public void onCancel(DialogInterface dialog) {
+            getActivity().getSharedPreferences("doze_settings", Activity.MODE_PRIVATE)
+                    .edit()
+                    .putBoolean("first_help_shown", true)
+                    .commit();
+        }
+    }
+
+    private void showHelp() {
+        HelpDialogFragment fragment = new HelpDialogFragment();
+        fragment.show(getFragmentManager(), "help_dialog");
+    }
+}
